+++ minijinja/src/defaults.rs
@@ -79,6 +79,8 @@ pub(crate) fn get_builtin_filters() -> BTreeMap<Cow<'static, str>, filters::Boxe
         rv.insert("indent".into(), BoxedFilter::new(filters::indent));
         rv.insert("select".into(), BoxedFilter::new(filters::select));
         rv.insert("reject".into(), BoxedFilter::new(filters::reject));
+        rv.insert("selectattr".into(), BoxedFilter::new(filters::selectattr));
+        rv.insert("rejectattr".into(), BoxedFilter::new(filters::rejectattr));
         rv.insert("map".into(), BoxedFilter::new(filters::map));
         #[cfg(feature = "json")]
         {